revert to old mode

+++ src/assets/node.sh
@@ -11,13 +11,17 @@ echo "Major: $MAJOR_NODE_VERSION"
 echo "Minor: $MINOR_NODE_VERSION"
 echo "Patch: $PATCH_NODE_VERSION"
 
-unset NVM_DIR
 export NVM_DIR="/.nvm"
 # Install nvm
-curl -o- https://raw.githubusercontent.com/creationix/nvm/v0.33.6/install.sh | bash
-[ -s "$NVM_DIR/nvm.sh" ] && . "$NVM_DIR/nvm.sh"
+if [ -s "$NVM_DIR/nvm.sh" ]; then
+  echo "nvm is already installed. Sourcing nvm..."
+  [ -s "$NVM_DIR/nvm.sh" ] && . "$NVM_DIR/nvm.sh"
+else
+  echo "nvm not found. Installing"
+  curl -o- https://raw.githubusercontent.com/creationix/nvm/v0.33.6/install.sh | bash  
+  [ -s "$NVM_DIR/nvm.sh" ] && . "$NVM_DIR/nvm.sh"
+fi
 
-# Install custom Node.js version if it matches specific criteria
 if [[ $MAJOR_NODE_VERSION == "14" && $MINOR_NODE_VERSION -ge 21 && $PATCH_NODE_VERSION -ge 4 ]]; then
   NODE_INSTALL_PATH="/.nvm/versions/node/v$NODE_VERSION"
 
@@ -30,55 +34,31 @@ if [[ $MAJOR_NODE_VERSION == "14" && $MINOR_NODE_VERSION -ge 21 && $PATCH_NODE_V
     curl "https://static.meteor.com/dev-bundle-node-os/v$NODE_VERSION/node-v$NODE_VERSION-linux-x64.tar.gz" | tar xzf - -C /tmp/
     mv /tmp/node-v$NODE_VERSION-linux-x64 $NODE_INSTALL_PATH
   fi
-
-  # Use the custom Node.js version
-  nvm use "$NODE_VERSION"
-  nvm alias default "$NODE_VERSION"
-  #export PATH="$NODE_INSTALL_PATH/bin:$PATH"
 else
-  # Fall back to nvm for other Node.js versions
-  echo "Installing and using Node.js via NVM..."
-  nvm install "$NODE_VERSION"
-  nvm use "$NODE_VERSION"
-  nvm alias default "$NODE_VERSION"
-  #export PATH="$NVM_DIR/versions/node/v$NODE_VERSION/bin:$PATH"
+  echo "Using NVM"
+  nvm install $NODE_VERSION
 fi
-echo "Current PATH: $PATH"
 
 # Verify installed Node.js and npm versions
 echo "Node.js version: $(node -v)"
 echo "NPM version before update: $(npm -v)"
 
+nvm use $NODE_VERSION
+nvm alias default $NODE_VERSION
+
 # Update npm to the specified version
 echo "Updating npm to version $NPM_VERSION..."
-npm install -g npm@"$NPM_VERSION"
+npm i -g npm@$NPM_VERSION
 echo "NPM version after update: $(npm -v)"
-#export NODE_PATH=$(dirname $(nvm which $(node --version)))
+export NODE_PATH=$(dirname $(nvm which $(node --version)))
+
+APP_PATH="$(/opt/elasticbeanstalk/bin/get-config container -k app_staging_dir)"
+echo "APP_PATH: $APP_PATH"
 
 # AWS Linux 2 /2023
 [[ -z "$APP_PATH" ]] && APP_PATH="$(/opt/elasticbeanstalk/bin/get-config platformconfig -k AppStagingDir)"
 echo "APP_PATH: $APP_PATH"
 
-if [ ! -d "$APP_PATH" ]; then
-  echo "Error: App staging directory does not exist. Probably a config update Dependencies should be there already"
-  CURRENT_PATH="$(echo $(/opt/elasticbeanstalk/bin/get-config platformconfig -k AppDeployDir))"
-  ls "$CURRENT_PATH"
-  exit 0
-else
-  cd "$APP_PATH"
-fi
-ls "$APP_PATH"
-
-echo "Installing root dependencies..."
-npm config delete prefix
-npm install
-if [ -d "programs/server" ]; then
-  cd programs/server
-  echo "Installing Meteor server dependencies..."
-  npm install --unsafe-perm
-else
-  echo "Error: Meteor server directory not found."
-  exit 1
-fi
-
-
+cd "$APP_PATH"
+ls
+cd programs/server && npm install --unsafe-perm
